made version be of type long

+++ logdb/src/main/java/org/logdb/Config.java
@@ -2,5 +2,5 @@ package org.logdb;
 
 public final class Config
 {
-    public static final int LOG_DB_VERSION = 1;
+    public static final long LOG_DB_VERSION = 1;
 }
